requirements.txt
django==4.1.7
Pillow==9.5.0
dj-database-url==1.2.0
psycopg2==2.9.5
python-decouple==3.7
djangorestframework==3.14.0
drf-yasg==1.21.5
djoser==2.2.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0